should be long ...
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@11400 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/php/phpbbs_post.c
@@ -387,7 +387,7 @@ PHP_FUNCTION(bbs_postarticle)
     FILE *fp;
     int NBUser = 0;
 #ifdef NEWSMTH
-    int from = 0;
+    long from = 0;
 #endif
     char name[STRLEN];
